add tree algorithm
